Projectile Collision
Including switch triggers and tile collision in general. Projectile
constructors were modified to allow any instance to either trigger an
event, and what to do on collision. For example, a thrown projectile
doesn't die when colliding, instead it stops moving and begins its
attack.

The event type can also be passed in to the projectile.